chore: Explicitly check if nimbus-cli has been downloaded and installed

+++ scripts/validate-nimbus-file.sh
@@ -9,10 +9,18 @@ if [ -d "${TARGET_PATH}" ]; then
 else
   echo "Create directory for \`nimbus-cli\`: $TARGET_PATH"
   mkdir -p $TARGET_PATH
+fi
 
+if [ -f "${NIMBUS_INSTALL_LOCAL_PATH}" ]; then
+  echo "Skip download of \`nimbus-cli\` installer"
+else
   echo "Download \`nimbus-cli\` installer from $NIMBUS_INSTALL_REMOTE_PATH to $TARGET_PATH"
   curl --proto '=https' --tlsv1.2 -sSf -o "$NIMBUS_INSTALL_LOCAL_PATH" "$NIMBUS_INSTALL_REMOTE_PATH"
+fi
 
+if command -v "${TARGET_PATH}/nimbus-cli"; then
+  echo "Skip installation of \`nimbus-cli\`"
+else
   echo "Install \`nimbus-cli\`"
   bash $NIMBUS_INSTALL_LOCAL_PATH --directory $TARGET_PATH
 fi
